Extract dynamic admission types
