+++ crates/shelterwood/src/tree.rs
@@ -15,15 +15,13 @@ use crate::{
     ActorDef, ActorOnceDef, ActorRef, ChildId, DefaultsInheritance, Intensity, LifecycleEvents,
     Membership, ReadinessDeadline, RestartPolicy, Retention, ScopeDefaults, ScopeSnapshot,
     Shutdown, ShutdownTimeout, SnapshotReceiver, Strategy, WaitError,
+    admission::{RemoveOutcome, ReserveError},
     cells::{MemberStage, ScopeCell},
     definition::DefinitionSource,
     driver::DynamicReservation,
     exit::{StartupError, StopReason},
     mailbox::MailboxCell,
-    plan::{
-        BuilderCore, ChildConstruction, LowerError, RemoveOutcome, ReserveError, ScopeConstruction,
-        SlotCell,
-    },
+    plan::{BuilderCore, ChildConstruction, LowerError, ScopeConstruction, SlotCell},
     policy::{CommonOptions, InvalidPolicy, ResolvedDefaults, ScopeFlavor},
     raw::{RawDef, RawOnceDef},
     runtime::{self, Latch},
