+++ tools/enforcement-fixtures/skeleton/crates/shelterwood/src/admission.rs
@@ -1,0 +1,2 @@
+// Placeholder mirroring crates/shelterwood/src/admission.rs so every path the
+// enforcement scripts scan exists in the fixture tree.
